Fix build warnings on iOS/tvOS

+++ src/image.h
@@ -60,7 +60,7 @@ public:
 	}
 
 	int NumColumns() const { return m_columns; }
-	int NumRows() const { return m_data.size() / m_columns; }
+	int NumRows() const { return int(m_data.size() / m_columns); }
 
 	void AddRow(const std::vector<double> &row)
 	{
